+++ CHANGELOG.md
@@ -8,3 +8,10 @@
 
 - 2023-11-23: v0.1.0
   - **Fix**: Height of formula will adapt to the font size of your editor now.
+
+- 2024-01-07: v0.2.0
+  - **Feature**: 
+    1. Use it in language that you want.
+    2. Multi-line formula is supported.
+    3. Style of formula code is supported.
+    4. Set the symbol other than `$` to mark formula (just support '@' or '#' now).